% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{create_vaja}
\alias{create_vaja}
\title{Basic functions}
\usage{
create_vaja(
  age,
  weight,
  id,
  alive = TRUE,
  year = 0,
  wcalf = FALSE,
  slaughtered = FALSE
)
}
\description{
Basic functions
}
