% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individ_Stoch_Sims_weight_dep.R
\name{SimulateHerd}
\alias{SimulateHerd}
\title{Function to simulate a reindeer herd}
\usage{
SimulateHerd(
  calving.rate = c(0.05, 0.6, 0.75, 0.85, 0.95, 0.97, 0.97, 0.95, 0.9, 0.9, 0.9),
  winter.survival = rep(0.96, 10),
  winter.survival_calves = 0.85,
  summer.survival = rep(0.99, 11),
  summer.survival_calves = 0.96,
  autumn.survival = rep(0.99, 11),
  autumn.survival_calves = 0.96,
  slaughter.adults = c(0.01, 0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 1),
  mean.weight = c(40, 60, 70, 75, rep(80, 5), 77, 75, 75),
  slaughter.calves = 0.56,
  standard_vaj_weight = 75,
  b_weight = 0.25,
  b_calving = 0.01,
  sd.repeatability = 2,
  weight.dependent_calving_probability = TRUE,
  weight.dependent_calf_sluaghter = TRUE,
  n_years = 50,
  plot_output = FALSE
)
}
\arguments{
\item{calving.rate}{The proportion of females calving per age class}

\item{winter.survival}{The probability of winter survival for non-calf females per age class}

\item{winter.survival_calves}{The probability of winter survival for calves}

\item{summer.survival}{The probability of summer survival for non-calf females per age class}

\item{summer.survival_calves}{The probability of summer survival for calves}

\item{autumn.survival}{The probability of autumn survival for non-calf females per age class}

\item{autumn.survival_calves}{The probability of autumn survival for calves}

\item{slaughter.adults}{Proportion slaughtered females per age class}

\item{mean.weight}{Expected mean weight (kg) per age class}

\item{slaughter.calves}{Proportion slaughtered females calves.}

\item{standard_vaj_weight}{Standard weight for an adult female that the weight dependence has as baseline.}

\item{b_weight}{Regression parameter relating a calf's body weight to its mother's body weight}

\item{b_calving}{Regression parameter relating a female's probability of calving to her body weight}

\item{sd.repeatability}{The amount of noise added to the body weight each year}

\item{weight.dependent_calving_probability}{Specifies if the reproduction and calf weight should depend on the mother's weight}

\item{weight.dependent_calf_sluaghter}{Specifies if selection on calf weights is applied or not}

\item{n_years}{The number of years to simulate. Default is 50.}

\item{plot_output}{Specifies whether results should be plotted every year}
}
\value{
A list consisting of 7 values for each individual: age, weight, id, alive, year, wcalf, slaughtered.
}
\description{
This function simulates the female part of a herd for 50 years
}
\details{
The default input values correspond to a herd in very good condition and hardly any predation.
The parameter slaughter.calves is used to tune the growth of the herd.
The following parameters are used to model weight dependence: standard_vaj_weight, b_weight, and b_calving.
The parameter standard_vaj_weight is used in the calving() and create_calves() functions
The parameter b_weight is used in the create_calves() function
The parameter b_calving is used in the calving() function. The relationship is given by a logit function.
The parameter sd.repeatability is used in the aging() function
}
\examples{
Herd1 <- SimulateHerd()

}
